Add in-code documentation and update examples (#10)
## Contents

1. Adds new `helloworld` example app with lots of documentation. This
example will be used by the CLI in a future PR instead of embedding the
example in the CLI binary.
2. Updates `opentofu` example app with documentation and more methods. I
also expanded the example so it will work across different projects.
This uses a "stateless" approach to operating. I chose to omit the List
function as it doesn't really work for OpenTofu unless the provider has
a specific datasource to list all of the resources of that type.